Fix rubocop offenses
[664]
